# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chainsaw Manual Training App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quizzes/{moduleId}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz questions for a module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
        - name: deviceId
          in: header
          required: true
          schema:
            type: string
        - name: activationCode
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quiz questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{moduleId}/submit:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      summary: Submit quiz answers
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
...
  /admin/news:
    post:
      operationId: createNewsItem
      tags: [news]
      summary: Create a news item
      parameters:
        - name: adminToken
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNewsItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
  /admin/news/pending:
    get:
      operationId: listPendingNewsItems
      tags: [news]
      summary: List news items pending admin approval
      parameters:
        - name: adminToken
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pending news items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
  /admin/news/fetch-now:
    post:
      operationId: triggerNewsFetch
      tags: [news]
      summary: Manually trigger an RSS fetch
      parameters:
        - name: adminToken
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Fetch result summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsFetchResult"
  /admin/news/retag-all:
    post:
      operationId: retagAllNewsItems
      tags: [news]
      summary: Retag all approved articles that have no LO/AC assigned
      parameters:
        - name: adminToken
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Retag result summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsRetagResult"
  /admin/news/{id}/approve:
    post:
      operationId: approveNewsItem
      tags: [news]
      summary: Approve a pending news item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: adminToken
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
  /admin/news/{id}/reject:
    post:
      operationId: rejectNewsItem
      tags: [news]
      summary: Reject (delete) a pending news item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: adminToken
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected
  /admin/news/{id}:
    patch:
      operationId: updateNewsItem
      tags: [news]
      summary: Update a news item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: adminToken
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNewsItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
    delete:
      operationId: deleteNewsItem
      tags: [news]
      summary: Delete a news item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: adminToken
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
components:
  schemas:
...
    Quiz:
      type: object
      required: [moduleId, moduleTitle, questions]
      properties:
        moduleId:
          type: integer
        moduleTitle:
          type: string
        passingScore:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
    QuizQuestion:
      type: object
      required: [id, question, options]
      properties:
        id:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
    QuizSubmission:
      type: object
      required: [deviceId, activationCode, answers]
      properties:
        deviceId:
          type: string
        activationCode:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
    QuizAnswer:
      type: object
      required: [questionId, selectedOption]
      properties:
        questionId:
          type: integer
        selectedOption:
          type: integer
    QuizResult:
      type: object
      required: [passed, score, passingScore, correct, total]
      properties:
        passed:
          type: boolean
        score:
          type: integer
        passingScore:
          type: integer
        correct:
          type: integer
        total:
          type: integer
        feedback:
          type: array
          items:
            $ref: "#/components/schemas/QuizFeedback"
    QuizFeedback:
      type: object
      required: [questionId, correct, correctOption]
      properties:
        questionId:
          type: integer
        correct:
          type: boolean
        correctOption:
          type: integer
...
    NewsItem:
      type: object
      required: [id, title, excerpt, url, publishedAt, createdAt, status]
      properties:
        id:
          type: integer
        title:
          type: string
        excerpt:
          type: string
        url:
          type: string
        imageUrl:
          type: ["string", "null"]
        publishedAt:
          type: string
        createdAt:
          type: string
        status:
          type: string
        feedSource:
          type: ["string", "null"]
        learningOutcome:
          type: ["string", "null"]
        assessmentCriteria:
          type: ["string", "null"]
    NewsRetagResult:
      type: object
      required: [total, tagged, errors]
      properties:
        total:
          type: integer
        tagged:
          type: integer
        errors:
          type: integer
    NewsFetchResult:
      type: object
      required: [fetched, inserted, skipped, errors]
      properties:
        fetched:
          type: integer
        inserted:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
    CreateNewsItemInput:
      type: object
      required: [title, excerpt, url, publishedAt]
      properties:
        title:
          type: string
        excerpt:
          type: string
        url:
          type: string
        imageUrl:
          type: string
        publishedAt:
          type: string
    UpdateNewsItemInput:
      type: object
      properties:
        title:
          type: string
        excerpt:
          type: string
        url:
          type: string
        imageUrl:
          type: ["string", "null"]
        publishedAt:
          type: string
...